removing player from use menu
although the function we're calling promises to only show th evisible
objects and the player object is invisible,  player is still returned in
th elist.

if you try to use something with the player, the engine forbids it,
because the player is not visible. so there must be a bug in the engine.